Fix mpirun issue to run solvers in parallel
precice seems to be at the root of our issue with not being able to fork mpirun within gymprecice.

The main reason is that, within the precice Python-binding, developers have added  line
"from mpi4py import MPI" to deal with a dependency compile error (as a workaround), which
is still an open issue:
The issue: https://github.com/precice/python-bindings/issues/8#issuecomment-575329609
The PR which has not been merged into the release branch:
https://github.com/precice/python-bindings/pull/101
and
https://github.com/precice/precice/pull/299

Based on https://stackoverflow.com/questions/64581965/run-mpi-program-using-subprocess-popen,
the added line, "from mpi4py import MPI", makes the python script to be run in singleton mode,
and therefore:
sunprocess.run(["mpirun"], ...) fails
sunprocess.Popen(["mpirun"], ...) is zombified.

This fix is a workaround to remove MPI from the parent env  and provide the "cleaned" env to
subprocess rather than letting the subprocess inherit the parent env as it is.
The fix is inspired by the discussion in:
https://stackoverflow.com/questions/30384649/python-with-embedded-call-to-mpirun